Restructured the files for Image component and revamped the widget controls..
