+++ frontend/src/Editor/Components/Image/image.scss
@@ -1,0 +1,50 @@
+.broken-url-placeholder {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 16px;
+    gap: 10px;
+    background-color: #FFFFFF;
+
+    p {
+        text-align: center;
+        color: #11181C;
+        font-family: "IBM Plex Sans";
+        font-size: 14px;
+        font-weight: 500;
+        line-height: 20px;
+    }
+}
+
+.img-control-wrapper {
+    position: fixed;
+    left: 16px;
+    bottom: 16px;
+    padding: 2px;
+    display: flex;
+    align-items: center;
+    gap: 2px;
+    border-radius: 6px;
+    background: #FFFFFF;
+    box-shadow: 0px 0px 1px 0px #3032330D, 0px 2px 4px 0px #3032331A;
+    opacity: 0;
+    visibility: hidden;
+
+    .img-control-btn {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        padding: 4px;
+        border-radius: 4px;
+        border: none;
+        background-color: #FFFFFF;
+    }
+}
+
+.image-widget-wrapper:hover .img-control-wrapper {
+    opacity: 1;
+    visibility: visible;
+}
