--- frontend/src/Editor/Components/image.scss
@@ -1,19 +1,0 @@
-.broken-url-placeholder {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 16px;
-    gap: 10px;
-
-    p {
-        text-align: center;
-        color: #11181C;
-        font-family: "IBM Plex Sans";
-        font-size: 14px;
-        font-weight: 500;
-        line-height: 20px;
-    }
-}
